requirements.txt
aiosqlite==0.19.0
anyio==3.7.0
black==23.3.0
certifi==2023.5.7
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.99.1
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
identify==2.5.24
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pluggy==1.2.0
pre-commit==3.3.3
psycopg==3.1.9
psycopg-binary==3.1.9
pydantic==1.10.10
pytest==7.4.0
pytest-asyncio==0.21.0
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.17
starlette==0.27.0
typing_extensions==4.7.1
virtualenv==20.23.1
